// src/api/reports/enums/mil-finding-taxonomy.enum.ts
export enum MilFindingTaxonomy {
  ARTIFICIAL_URGENCY = 'ARTIFICIAL_URGENCY',
  FEAR = 'FEAR',
  ANGER = 'ANGER',
  SOCIAL_PRESSURE = 'SOCIAL_PRESSURE',
  CALL_TO_VIRALITY = 'CALL_TO_VIRALITY',
  FALSE_AUTHORITY = 'FALSE_AUTHORITY',
  MISSING_DATE = 'MISSING_DATE',
  MISSING_LOCATION = 'MISSING_LOCATION',
  MISSING_BASELINE = 'MISSING_BASELINE',
  QUOTE_TRUNCATION = 'QUOTE_TRUNCATION',
  OLD_CONTENT_PRESENTED_AS_CURRENT = 'OLD_CONTENT_PRESENTED_AS_CURRENT',
  HEADLINE_MISMATCH = 'HEADLINE_MISMATCH',
  SOURCE_IMPERSONATION = 'SOURCE_IMPERSONATION',
  LACK_OF_INDEPENDENT_EVIDENCE = 'LACK_OF_INDEPENDENT_EVIDENCE',
  UNSUPPORTED_STATISTICS = 'UNSUPPORTED_STATISTICS',
  AI_MEDIA_UNCERTAINTY = 'AI_MEDIA_UNCERTAINTY',
  SCREENSHOT_CONTEXT_LOSS = 'SCREENSHOT_CONTEXT_LOSS',
  AUDIO_UNCERTAINTY = 'AUDIO_UNCERTAINTY',
  VIDEO_CONTEXT_MISMATCH = 'VIDEO_CONTEXT_MISMATCH',
  EVIDENCE_EVALUATION = 'EVIDENCE_EVALUATION',
}